feat: user CRUD

+++ myproject/blog/views.py
@@ -1,8 +1,11 @@
-from django.shortcuts import render, get_object_or_404, redirect
-from .models import Post, Category
-from .forms import PostForm
+from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from django.core.paginator import Paginator
-from django.db.models import Q
+from .forms import PostForm, CreateUserForm, CustomAuthenticationForm
+from .models import Post, Category
 
 def home(request):
     query = request.GET.get('q')
@@ -13,7 +16,7 @@ def home(request):
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     categories = Category.objects.all()
-    return render(request, 'blog/home.html', {'posts': page_obj, 'query': query, 'categories': categories})
+    return render(request, 'home.html', {'posts': page_obj, 'query': query, 'categories': categories})
 
 def posts_by_category(request, category_id):
     category = get_object_or_404(Category, id=category_id)
@@ -22,36 +25,89 @@ def posts_by_category(request, category_id):
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     categories = Category.objects.all()
-    return render(request, 'blog/posts_by_category.html', {'posts': page_obj, 'category': category, 'categories': categories})
+    return render(request, 'posts_by_category.html', {'posts': page_obj, 'category': category, 'categories': categories})
 
+def register(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    else:
+        form = CreateUserForm()
+        if request.method == 'POST':
+            form = CreateUserForm(request.POST)
+            if form.is_valid():
+                form.save()
+                user = form.cleaned_data.get('username')
+                messages.success(request, 'Account was created for ' + user)
+                return redirect('login')
+
+        context = {'form': form}
+        return render(request, 'register.html', context)
+    
 def post_detail(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    return render(request, 'blog/post_detail.html', {'post': post})
+    is_author = post.author == request.user
+    return render(request, 'post_detail.html', {'post': post, 'is_author': is_author})
 
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    else:
+        if request.method == 'POST':
+            username = request.POST.get('username')
+            password = request.POST.get('password')
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                login(request, user)
+                return redirect('home')
+            else:
+                messages.info(request, 'Username OR password is incorrect')
+
+        form = CustomAuthenticationForm() 
+        context = {'form': form}
+        return render(request, 'login.html', context)
+
+@login_required(login_url='login')
+def logoutUser(request):
+    logout(request)
+    return redirect('login')
+
+@login_required(login_url='login')
 def post_create(request):
     if request.method == 'POST':
-        form = PostForm(request.POST, request.FILES)
+        form = PostForm(request.POST)
         if form.is_valid():
-            form.save()
+            post = form.save(commit=False)
+            post.author = request.user
+            post.save()
             return redirect('home')
     else:
         form = PostForm()
-    return render(request, 'blog/post_form.html', {'form': form})
+    return render(request, 'post_form.html', {'form': form})
 
-def post_update(request, pk):
+@login_required
+def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    if request.method == 'POST':
-        form = PostForm(request.POST, request.FILES, instance=post)
+    if post.author != request.user:
+        return redirect('home')
+    
+    if request.method == "POST":
+        form = PostForm(request.POST, instance=post)
         if form.is_valid():
             form.save()
-            return redirect('home')
+            return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm(instance=post)
-    return render(request, 'blog/post_form.html', {'form': form})
+    
+    return render(request, 'post_form.html', {'form': form})
 
+@login_required
 def post_delete(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    if request.method == 'POST':
+    if post.author != request.user:
+        return redirect('home')
+    
+    if request.method == "POST":
         post.delete()
         return redirect('home')
-    return render(request, 'blog/post_confirm_delete.html', {'post': post})
+    
+    return render(request, 'post_confirm_delete.html', {'post': post})
